Kokkos: Added Lambda variant of parallel_reduce(Iter, Functor, ReturnValue);
Added both the flat parallelism as well as the ParallelWorkRequest version
of parallel_reduce which supports Lambdas. This right now works only for
the variant where the ReturnValue is provided by reference.
As device_type Kokkos::DefaultDeviceType is used.
For init the reduction value is initialized as:

value = ValueType();

The join is performed as:

update += input;

The same code is used if a functor is provided without a device_type defined.

An example:
    double result = 0.0;
    Kokkos::parallel_reduce(Kokkos::ParallelWorkRequest(25,4),[=](const DeviceType& dev, double& sum)  {
      int i = dev.league_rank()*dev.team_size() + dev.team_rank();
      sum += a(i,1) + a(i,2);
      sum += a(i,0) - a(i,3);
      sum += a(i,4) + a(i,0);
      sum += a(i,2) - a(i,1);
      sum += a(i,3) + a(i,4);
    },result);

    double result = 0.0;
    Kokkos::parallel_reduce(100,[=](const int &i, double& sum)  {
      sum += a(i,1) + a(i,2);
      sum += a(i,0) - a(i,3);
      sum += a(i,4) + a(i,0);
      sum += a(i,2) - a(i,1);
      sum += a(i,3) + a(i,4);
    },result);

This code does not work with the result being a pointer.