Kokkos: Added Makefile support and Tutorial files
The Makefile.kokkos can be included by application Makefiles to compile
Kokkos object files as well as set the necessary defines. Settigns can
be given via command line or in the application Makefile prior to include
Makefile.kokkos. From the Makefile:

 This Makefile is intended to be include in an application Makefile.
 It will append the OBJ variable with objects which need to be build for Kokkos.
 It also will produce a KOKKOS_INC and a KOKKOS_LIB variable which must be
 appended to the compile and link flags of the application Makefile.
 Note that you cannot compile and link at the same time!
 If you want to include dependencies (i.e. trigger a rebuild of the application
 object files when Kokkos files change, you can include KOKKOS_HEADERS in your
 dependency list.
 The Makefile uses a number of variables which can be set on the commandline, or
 in the application Makefile prior to including this Makefile. These options set
 certain build options and are explained in the following.

The tutorial examples are intended as a generic introduction to Kokkos starting
with very low entry barrier examples and expanding to more complex issues.