Epetra: Added the fifth of five lessons.
I am translating the Tpetra lessons into Epetra, in preparation for
the VECPAR tutorial at the end of this month.  This commit adds the
fifth of the five lessons.

I have not yet plugged the lessons into Doxygen.  They build and pass
tests, but do not generate Doxygen documentation yet.  This will
change in near-future commits.