Tribits: Fix issue with tarball creation
The recent updates to Tribits fixed a minor issue that the tarball exclude files
code had previously been working around. Previously there was a "/." at the end
of <project name>_SOURCE_DIR which interfered with excluding files since the
exludes are a regular expression and not a path. That "/." has been removed and
the workaround previously in place was stripping meaningful information from
<project name>_SOURCE_DIR. The fix was to remove the work around. This caused
one conditional to be removed that I couldn't test well since it was for the
case when Trilinos is an external repo to another project. Both paths of that
conditional became the same code so I don't believe this will cause an issue.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=136,notpassed=0 (0.84 min)
1) SERIAL_RELEASE => passed: passed=136,notpassed=0 (0.71 min)