TrilinosCouplings:  Make MeanBasePreconditioner available to all scalars
I reworked the MeanBasedPreconditioner to work with all scalar types,
only requiring a specialization of the GetMeanValsFunc for each scalar
type.  In particular I set it up so that it works with the ensemble
scalar type based on averaging the values across the ensemble.  For
"double", it essentially just creates a copy of the matrix.  This allows
us to at least explore the effectiveness of the mean-based
preconditioner for the ensemble approach, even though it isn't an
efficient implementation.