MueLu: fixing failing tests on Mac
Long story short: regular expressions in diff -I work differently on
Mac and Linux. If the last line of file does not have a new line, ^.*$
on Linux includes the whole line (i.e., $ is the end of file), while on
Mac it does not.

The missing end of line originated in Level, where printing of a level
resulting in missing end of line if the verbosity level did not include
Debug.

Build/Test Cases Summary
Enabled Packages: MueLu
Enabled all Forward Packages
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=189,notpassed=0 (23.34 min)