MueLu: update adapters to create the proper sized nullspace for easy XML
CreatePreconditioner routines construct the nullspace if a user does not
provide it, but only check for the old style XML interface. I've added a
check for the new one. This way, if you don't provide the new style, it
would generate the nullspace of size of the "number of equations".

Build/Test Cases Summary
Enabled Packages: MueLu
Enabled all Forward Packages
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=189,notpassed=0 (11.99 min)