MueLu: created master list of easy MueLu parameters
The idea is to have a single XML file with all parameters, and use XSLT
to transform it to our needs:
- Use tex.xsl to transform it to a LaTeX input file, which goes in users
  guide
- Use html.xsl to transform it to the web documentation
- Use paramlist.xsl to transform it into an hpp file included into
  parameter list interpreter

Now, adding a new parameter in User API is done in a simple way:
1. Add a <parameter> entry with name, type, default value (optional) and
   description to masterList.xml
2. Run update_params.sh
   This updates:
   - paramlist.tex which is included in the mueluguide.tex
   - src/MueCentral/MueLu_MasterList.cpp

Another cool feature is that because MasterList class is a singleton and
all functions are static, we can get the default values in factories
from this list. So, for instance, to set a smoothed aggregation factory
damping parameter, we set the default value in the masterList.xml, and
call MasterList::getDefault<double>("sa: damping factor") in the
factory. This way we will avoid any inconsistencies between default values in
factories, and default values in the user documentation (once the
factories have been converted).

Build/Test Cases Summary
Enabled Packages: MueLu
Enabled all Forward Packages
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=189,notpassed=0 (31.75 min)