Setting up to use default email git property 'mailinglist' if dirs_to_emails is missing
With this change, if the file hooks/dirs_to_emails is missing, then it
will instead use the git config property 'mailinglist'.

This is an important change because it will allow the usage of git
config [hooks] property 'mailinglist'.  This property can be directly
set in the gitolite-admin/confi/gitolite.conf file.  This allows
anywone who can push to the gitolite-admin repo to be able to able to
change the list of email addresses that will be pushed to.