Kokkos: Added "Advanced_Views" tutorial examples to CMake build system.
In my previous commit, I added some of the Kokkos tutorial examples to
the CMake build system, so that they will build in the check-in and
nightly tests.  This commit adds all of the tutorial examples in the
Advanced_Views subdirectory.  The example that exercises CUDA UVM
builds conditionally on the CMake option Kokkos_ENABLE_Cuda_UVM.  I
also fixed build warnings in some tutorial examples.