Tpetra: Distributor: Fixed other Bug 6170 - related issues.
These don't manifest in the Bug 6170 test, but could pop up in other
tests with the same Maps, but with CrsGraph or CrsMatrix instead of
MultiVector.

I also took the opportunity to make Distributor's error messages more
informative and consistent.

I'm forcing the push because of a heisenbug in the "classic" version
of Tpetra, with MultiVector unit tests, that has nothing to do with
Bug 6170 or its fixes.

Build/Test Cases Summary
Enabled Packages: KokkosCore, KokkosCompat, KokkosLinAlg, KokkosContainers, KokkosMpiComm, Tpetra
0) MPI_DEBUG => passed: passed=72,notpassed=0 (1.50 min)
1) SERIAL_RELEASE => passed: passed=52,notpassed=0 (0.05 min)
2) MPI_RELEASE_KOKKOS_REFACTOR => passed: passed=84,notpassed=0 (1.45 min)
Other local commits for this build/test group: 6a23456, 909de40