[Sundance] prefix pdeopt, playa with "sundance"
Up until now, all Sundance libraries are prefixed with "sundance", except
pdeopt and playa. This is not only inconsistent, but also causes trouble since
the library name "pdeopt" is very generic and is bound to collide with other
libraries.
This commit brings consistency to the sundance naming scheme by prefixing
pdeopt and playa with "sundance" as well.