Tpetra: Improved CMake error message for missing Kokkos subpackages.
Enabling the Kokkos refactor version of Tpetra requires that certain
Kokkos subpackages be enabled.  These packages are not enabled by
default.  Thus, users must go through some extra effort to get the
Kokkos refactor version of Tpetra to build.  It's not immediately
obvious how to do this.

This commit improves the error message which Tpetra's CMake
configuration displays if Kokkos refactor is enabled, but the required
Kokkos subpackages are not.  The error message used to mention which
subpackages were required.  The improved message now also explains how
to enable those subpackages.  Thus, the error message now should
suffice for showing users how to enable the Kokkos refactor version of
Tpetra.