Xpetra: Made CrsMatrix unit test more thorough, to catch a possible bug.
There is some issue, apparently related to
Tpetra::CrsMatrix::replaceLocalValues, that causes this Xpetra test
and one MueLu test (Maxwell 3-D) to fail.  I have enhanced the Xpetra
test in an attempt to find the source of the bug.  replaceLocalValues
appears to work fine, until we try to do a mat-vec.  I'm not sure
what's going on.  Note that the test FAILS with Kokkos refactor
Tpetra, but does not fail with "classic" Tpetra (currently the
default).