Disabling STK in automated testing
STK has been broken in the build on the CASL Fissile machines for a long time,
since the update of STK from SIERRA.  Basically we are getting a lot of build
errors in STK being reported to CDash.

Also, I am hoping that disabling STK will speed up the build of Trilinos so
that we don't timeout the MPI_RELEASE_GCC build after 5 hours like is occuring
now.