Tpetra: Ported Xpetra CrsMatrix,Tpetra_ReplaceLocalValues test.
The Xpetra CrsMatrix, Tpetra_ReplaceLocalValues test fails with the
Kokkos refactor version of Tpetra.  In order to figure out whether
this is a Tpetra or Xpetra issue, I ported the test to Tpetra (e.g.,
by replacing "Xpetra::" with "Tpetra::", and using constructors
directly instead of through factories).  It still fails in Tpetra, so
we can focus on fixing Tpetra first.

The test also now verifies that apply() works when the matrix is
initially constructed as the identity matrix.  The issue is that the
resumeFill and modification doesn't work.

NOTE: This test currently FAILS with the Kokkos refactor version of
Tpetra.  It should pass with the "classic" version of Tpetra.

Build/Test Cases Summary
Enabled Packages: KokkosCore, KokkosCompat, KokkosContainers, KokkosLinAlg, KokkosMpiComm, Tpetra
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => FAILED: passed=74,notpassed=2 => Not ready to push! (2.50 min)
3) SERIAL_RELEASE_SS => FAILED: passed=53,notpassed=2 => Not ready to push! (0.94 min)
4) MPI_DEBUG_KOKKOS_REFACTOR => FAILED: passed=83,notpassed=2 => Not ready to push! (3.13 min)
5) MPI_RELEASE_KOKKOS_REFACTOR => FAILED: passed=74,notpassed=2 => Not ready to push! (2.29 min)
Other local commits for this build/test group: da68677
WARNING: Forced the push!