Kokkos: Added Atomic View capabilities.
The atomic view currently only works for 32bit and 64bit scalars.
It supports a (hopefully) comprehensive list of operators
(=,+=,/=,-=,++ .......) and should in principle work everywhere.
It does not currently support taking a reference to a value.
The internal implementation works via a temporary object
AtomicDataElement which gets returned by the view instead of
the value_type. This objects provides atomic access operators as
well as the conversion operators to the value_type. A user will never
see that object.