KokkosLinAlg: Added sequential sparse kernels header files.
We want to make it easier to finish the Kokkos refactor port of
Tpetra, by replacing the "classic Kokkos" components on which Tpetra
currently depends.  A big source of complexity in Tpetra::CrsMatrix is
the "local mat ops" or "local sparse ops" object, which comes from
CrsMatrix's fifth template parameter (and CrsGraph's fourth template
parameter).  The Kokkos refactor version of Tpetra really only needs
this object for implementations of sparse matrix computational kernels
that don't yet have thread-parallel implementations in new Kokkos.

This commit adds implementations of the missing computational kernels
to the KokkosLinAlg subpackage of new Kokkos.  This will help us get
rid of the "local sparse ops" object in the Kokkos refactor version of
Tpetra.  The implementations are sequential; they do not use threads.
In time, we will add thread-parallel implementations of these
computational kernels.