Stokhos:  Add elementWiseMultiply and scale kernel PCE specializations.
The element-wise multiply implementation is assuming a mean-based
structure where the vector being multiplied is constant.  This will
likely break things when this assumption isn't true, such as the full
PCE preconditioner.  But that approach isn't very effective at this
point, whereas the mean-based one is, so this is what we're going with
for now.  I should add some sort of warning/error message when this
assumption isn't satisfied though.