Tpetra: Fixed bugs in KokkosRefactor Map when compiling for CUDA
In particular this gets rid of lazy initialization of the device version
of the map. The latter resulted in the underlying storage views disappearing
at the ned of the constructor, because certain kokkos views went out of scope.
As a consequence (when compiling with UVM) the host views disappeared, because
they were unmanaged.