This commit improves some error reporting for mesh consistency checking.
Also revert create_edges to use the original version of modification_end temporarily.
The newer edge-creation-specific modification_end made dramatic improvements to
scalability, but encounters a bug in a new adaptivity case we are now debugging,
where the scenario appears to involve creation of edges on some elements but not
others (not creating edges on 'parent' elements, with an interface between those
different elements coinciding with a processor boundary.
This is a scenario that we didn't have when developing the faster modification_end functionality.

A following commit will add a unit-test replicating at least part of the scenario
that is tripping us up in the parallel adaptivity case.

(From Sierra)