SEACAS: Update Trilinos build to remove unnecessary dependencies
There were a number of places that the libraries inside of SEACAS were being
listed as dependencies of other libraries or applications. This kind of
dependency is automatically handled by Tribits and re-stating the dependency
directly can have consequences to certain build modes or if the library names
change later. This removes those redundant dependencies.

(From Sierra)