Tpetra: Crs{Graph,Matrix}: Removed legacy KokkosClassic rowPtrs_ array.
This commit only affects the Kokkos refactor version of Tpetra.

In the "classic" version of Tpetra, the rowPtrs_ array stores row
offsets (the "ptr" part of the "ptr", "ind", "val" arrays of
compressed sparse row storage) in CrsGraph.  rowPtrs_ stores row
offsets for possibly unpacked "1-D" storage before fillComplete.
After fillComplete, if the "Optimize Storage" parameter of
fillComplete is false, it continues to store unpacked row offsets.
Otherwise, it stores packed row offsets (the default behavior).

The rowPtrs_ array (a Teuchos::ArrayRCP<size_t>) is superfluous for
the Kokkos refactor version of Tpetra.  We had already made it just a
view of the new Kokkos data structure (k_rowPtrs_).  This commit
removes it completely from CrsGraph, and removes all references to it
in CrsMatrix.