MueLu: fixed an issue with UC Dirichlet points
The problem was that we called PreserveDirichletAggregationAlgorithm
optionally. That means that for other algorithms some of the points
could still have BOUNDARY status, which they did not expect. Now, we
always call the algorithm, but check the option to preserve inside:
- If we preserve Dirichlet nodes, we aggregate them into a singleton
  aggregate, and change the status to IGNORED.
- If we don't, we change the status to IGNORED.

IGNORED is important, and must be different from AGGREGATED, as
incoming connections into a Dirichlet node (for instance, a user
zeroes out only Dirichlet row but not column) are valid for AGGREGATED
(and point could be added to Dirichlet aggregates) but not valid for
IGNORED (which remains singleton).

Build/Test Cases Summary
Enabled Packages: MueLu
Enabled all Forward Packages
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=189,notpassed=0 (8.17 min)
Other local commits for this build/test group: 6aee2e7, e1393dd