Making GNUInstallDirs optional and check for CMake version (Trilinos #6122)
I chagned the logic to maintain the old TriBITS defaults but instead let each
TriBITS project decide if to use GNUInstallDirs by default or not.  For this
purpose, I have defined the cache variable ${PROJECT_NAME}_USE_GNUINSTALLDIR
with the default provided by ${PROJECT_NAME}_USE_GNUINSTALLDIR_DEFAULT.  If
not specified, TriBITS will set
${PROJECT_NAME}_USE_GNUINSTALLDIR_DEFAULT=FALSE.  That way, we don't break
backward compatibility for all TriBITS projects but we let each TriBITS
project decide on its own what the default should be and we let the user
override that default.  Who could ask for more than that?

Also, I put in an if staement for The CMake version so GNUInstallDirs will not
be used if the version is too old.

I have updated all documentation, including a warning about using PATH
datatype for relative install paths (watch out).

See Trilinos #6122 for more details.