Making GlobalMPISession constructor call std::terminate() not throw (Trilinos #6211)
Here I remove all of the calls to TEUCHOS_TEST_FOR_EXCEPTION_PRINT() which can
throw excpetions and instead just print the error message and call
std::terminate()

It is normally desired to throw exceptions in constructors if something goes
wrong. This however creates problems for GlobalMPISession() since we cannot
principally catch those (with the TEUCHOS_STANDARD_CATCH_STATEMENTS logic).

The GlobalMPISession class constructor is *only* to be used to construct an
object in main() (and before everything else that might want to use MPI). This
is a special class and a special constructor and we might consider just
calling abort() (or the modern C++ std::terminate()) if any of the MPI
functions in this constructor return non-zero error codes. After all, if you
can't initialize MPI you can't do anything in an MPI program (except print an
error message and exit).

NOTE: This should maintain 100% backward compatibility in that an uncaught
exception will call std::terminate().  Here we are just jumping to the same
comclusion is a way that does not flag an uncaught exception by static
analysis tools.

For more details, see Trilinos bug #6211.