Tpetra: CrsGraph: Update to fix for Bug 6215.
This commit affects both the "classic" and Kokkos refactor versions of
Tpetra.

This relates to Paul Lin's commit on Mon 01 Sep 2014:

417e155482c9b1673e757ca2509b7af3ab0ff905

which was a fix for Bug 6215.  This commit updates that bug fix by
reintroducing the original intended optimization of only keeping the
original GID ordering if requested by the user.  It also adds the bug
fix to the Kokkos refactor version of Tpetra (Paul's commit only
changed the "classic" version of Tpetra).