Kokkos: Fixed missing copy constructor in Random class, updated nvcc_wrapper
The update of the nvcc_wrapper fixes some issues which crept up when
trying to compile OpenMPI 1.8.2. It now works. The script also has now
defaults for the device architecture (-arch=sm_35) and the host compiler
(-ccbin g++). Both can be overwritten by providing -arch and
 -ccbin respectively. You also can set them at the top of the wrapper script.
This is for example useful if you want two scripts - one with g++ and one
with icpc as backend compiler. You then can use a MPI compiled with the script
without providing any nvcc specific arguments.