Adding <Project>_TRIBITS_PACKAGE_USE_TRIBITS_DIR to avoid CTest/CDash failures (Trilinos #6218)
The old logic pushed in the commit:

98a3fee "Using TriBITS from Trilinos_TRIBITS_DIR for the TriBITS package (Trilinos #6218)"
Author: Roscoe A. Bartlett <bartlettra@ornl.gov>
Date:   Sun Sep 7 11:17:52 2014 -0400 (27 hours ago)

resulted in the defintion of the TriBITS package dir to be
Trilinos_TRIBITS_DIR in all cases which is very bad when using the the
CTest/CDash driver in TribitsCTestDriverCore.cmake where Trilinos_TRIBITS_DIR
is set to a directory defined outside of the inner Trilinos source dir.  This
results in a very nicely formatted error message that tells you exactly that.

With this new logic, ${${PROJECT_NAME}_TRIBITS_DIR} is ony used to define the
TriBITS package when ${PROJECT_NAME}_TRIBITS_PACKAGE_USE_TRIBITS_DIR=TRUE.

This commit should fix all of the automated Trilinos testing.