Anasazi: Added new test for Tpetra MultiVecTraits::SetBlock.
The Tpetra specialization of MultiVecTraits now has a new test for
SetBlock, TpetraSetBlock4.  This test manifested the issue that Alicia
Klinvex and others discovered.

TODO: Port this test to Tpetra.  It makes an excellent test for the
postcondition of Tpetra::deep_copy, that it does not invalidate any
views of the output MultiVector dst.

TODO: Port this test to Belos (to exercise Belos' MultiVecTraits).