Tpetra::MultiVector: Fixed deep_copy.
This commit affects both the "classic" and Kokkos refactor versions of
Tpetra.

The implementation of Tpetra::deep_copy for MultiVector arguments had
a bug: it was invalidating existing views of the output argument dst.
This manifested as incorrect behavior of the Tpetra specializations of
{Anasazi, Belos}::MultiVecTraits::SetBlock.

This commit fixes this, by ensuring that the only thing that changes
in dst is the contents of its storage.  Its pointers don't change, it
doesn't reallocate memory, and it doesn't change any other fields.

The easiest and cleanest way for me to fix this was to add a public
'assign' method to MultiVector.  This takes a MultiVector of the same
type, and does a deep copy into *this.  This means that the
specialization of deep_copy for when src and dst have the same type
doesn't need an explicit instantiation; it's just a one-line call to
assign().  Of course, the "generic" version of deep_copy, for when src
and dst have different types, takes more work.