Kokkos: Fix issue with Trilinos specific install
Steven Hamilton reported an issue was reported where Kokkos was installing into
the wrong place when not built inside of Trilinos. The issue was that Kokkos was
relying on Trilinos project specific variable instead of using the more generic
versions. The fix was to switch to the generic versions provided by Tribits.