Add ability to use a pre-created ETI macro header file (PHI Kanban #2719)
When you have a lot of instantiations, this file takes a long time to
generate.  When you enable DataTransferKit, it takes 24s to create the file!

I am going to add a pre-created version of this file to the VERA soruce repo
and we will update it whenever we update Trilinos into VERA.