Gtest: adding symlink to gtest for a non-fused source header
STK introduced a new dependency on a gtest header that we do not have. It turns
out the version of gtest that is in Sierra is a non-fused source copy that
occasionally requires tests to include files other than gtest.h. The version
that is in Trilinos is a fused source copy and as such only has on header,
gtest.h. The solution was to create a symlink to gtest.h with the name that the
STK test is looking for. This will allow us to continue to have a fused source
copy of gtest and allow STK to operate with both the Sierra and Trilinos build
systems without code changes.