Add missing HAVE_BLAS, HAVE_LAPCK, HAVE_MPI defines
The AztecOO package has dependencies on BLAS, LAPACK, and MPI through Epetra
and yet does not set these vars before configuring AztecOO_config.h.in.  But
some other package in Trilinos sets HAVE_BLAS in the cache.  As a result, the
first time you configure AztecOO, HAVE_BLAS is *not* defined in the configured
in the created AztecOO_config.h file.  But the second time you configure it
does define HAVE_BLAS.  Therefore, if you build after the first configure,
when you reconfigure, it will trigger a complete rebuild of anything that
includes AztecOO_config.h.

The reason that I saw this behavior this is that when looking at updated files
in the build dir, this file was flagged as being updated after the first
configure.

The fix was trivial and just copies what was done for HAVE_FORTRAN_SUPPORT.