Kokkos: Added Kokkos::initialize(int [,int][,int]) for generic initialization
The new functions take a threads, device and numa-regions argument depending on
what the default execution space is.

Given a single argument it is interpreted as the number of threads, unless Cuda
is the default execution space, in which case it is interpreted as device id.
If two arguments are given the first is interpreted as number of threads and the
second as either number of numa regions or device id depending on whether Cuda
is the default executionspace. If three arguments are given, they are interpreted
as threads, numa regions, and device id respectivly.