Tpetra: MultiVector: Fixed bugs in multiply.
This commit affects both the "classic" and Kokkos refactor versions of
Tpetra.

MultiVector::multiply only knows how to do a local dense matrix-matrix
multiply for MultiVectors with constant stride.  Thus, if any of the
MultiVectors doesn't have constant stride, the method must make a deep
copy of it.  The deep copies were using the one-argument copy
constructor.  In the Kokkos refactor version of Tpetra, this always
does a _shallow_ copy.  In the "classic" version of Tpetra, this may
do a shallow or deep copy, depending on whether the MultiVector
instance has view semantics (see e.g., the getCopyOrView method).
Shallow copies are incorrect in this case: we need a deep copy, so
that the result has constant stride.  This commit changes those deep
copies to use the two-argument copy constructor, with Teuchos::Copy as
the second argument.  This _always_ does a deep copy.

Furthermore, in the Kokkos refactor version of MultiVector, multiply()
was not copying the result back into *this, if *this does not have
constant stride.  The "classic" version does this correctly.  This
commit changes multiply() in the Kokkos refactor version, to copy back
to *this if necessary.