Kokkos: added value enum to VerifyExec..; added MemorySpace::count(ptr)
The value in VerifyExecutionCanAccessMemorySpace makes certain
enable_if, if_c etc. pattern easier.
The MemorySpace::count(ptr) gives back the current reference count value
of the ptr (0 if its not tracked). This allows to do special destruction
operatations of objects which have a view of complex things.
I needed it because I have an object which contains a view of unmanged views
and I need to free the memory of the unmanaged views when the last
instance of the view of views goes away.