Kokkos: Added SegmentedView
SegmentedView is a meta view type which allocates memory on the fly (in a
parallel kernel). It does so by using segments. In effect its a view of views
which has appropriate access operators, dimension functions etc. defined.

The SegmentedView can be 1 to 8 dimensional and LayoutLeft and Right do work
(others are not yet tested). The segmentation happens on the first dimension.

When SegmentedView::grow(team,size) is called it allocates memory for another
segment, and adds it to the SegmentedView. The segment length (that is its
first dimension) needs to be a power of 2 and is the same for all segments.
When constructing the SegmentedView the maximum length must be given, but it
starts with length zero.

The grow function needs to be called with a team and every team_member has to enter
the function at the same time, with the same size value.

This thing works on Cuda. Performance is reasonable and just filling a vector with zeros
takes about twice as long as a normal view if the segment size is about 1M.

This commit also adds Tests. A tutorial example will be added in the future.